# src/web/docs/swagger.yaml
basePath: /provider-api/v1
definitions:
  v1error.ErrorResponse:
    properties:
      Code:
        type: integer
      Error:
        type: string
      StatusText:
        type: string
    type: object
  v1health.Health:
    properties:
      inbound_connections:
        items:
          type: InboundConnection
        type: array
      outbound_connections:
        items:
          type: OutboundConnection
        type: array
      service_name:
        type: string
      service_provider:
        type: string
      service_start_time_utc:
        type: string
      service_status:
        type: ServiceStatus
      service_version:
        type: string
      timestamp_utc:
        type: string
      uptime:
        type: number
    type: object
  v1product.Product:
    properties:
      category:
        type: ProductCategory
      description:
        type: string
      id:
        type: string
      in_stock:
        type: boolean
      manufacturer:
        type: string
      name:
        type: string
      price:
        type: string
    type: object
host: localhost:8001
info:
  contact: {}
  description: API Documentation
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: API Documentation
  version: "2.0"
paths:
  /health:
    get:
      consumes:
      - application/json
      description: It returns the health of the service
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1health.Health'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1error.ErrorResponse'
      summary: Get health of the service
      tags:
      - health
  /products/productID:
    get:
      consumes:
      - application/json
      description: It allows to reterive a Product by ID
      parameters:
      - description: Product ID
        in: path
        name: productID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1product.Product'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1error.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1error.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1error.ErrorResponse'
      summary: Get a Product
      tags:
      - product
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"